[TECH TASK]: Rename the project
