+++ settings.gradle.kts
@@ -1,1 +1,1 @@
-rootProject.name = "we-are-playing-basketball"
+rootProject.name = "basketball"
