+++ src/main/java/online/rabko/basketball/BasketballApplication.java
@@ -1,4 +1,4 @@
-package by.vibe.weareplayingbasketball;
+package online.rabko.basketball;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,7 +7,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * Main entry point for a Spring Boot application.
  */
 @SpringBootApplication
-public class WeArePlayingBasketballApplication {
+public class BasketballApplication {
 
     /**
      * Starts the Spring Boot application.
@@ -15,7 +15,7 @@ public class WeArePlayingBasketballApplication {
      * @param args command-line arguments
      */
     public static void main(String[] args) {
-        SpringApplication.run(WeArePlayingBasketballApplication.class, args);
+        SpringApplication.run(BasketballApplication.class, args);
     }
 }
