+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=we-playing-basketball
+spring.application.name=bastketball
 server.forward-headers-strategy=native
 logging.level.root=info
 springdoc.swagger-ui.urls[0].name=general
