+++ src/test/java/online/rabko/basketball/BasketballApplicationTests.java
@@ -1,10 +1,10 @@
-package by.vibe.weareplayingbasketball;
+package online.rabko.basketball;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class WeArePlayingBasketballApplicationTests {
+class BasketballApplicationTests {
 
     @Test
     void contextLoads() {
